complete redis codes

+++ src/main/java/com/eric/projects/bargainrush/redis/BasePrefix.java
@@ -14,8 +14,9 @@ public abstract class BasePrefix implements KeyPrefix {
         this.expireSeconds = expireSeconds;
         this.prefix = prefix;
     }
+
     @Override
-    public int expireSeconds() {
+    public long expireSeconds() {
         return expireSeconds;
     }
